+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- expose fields, vocabularies and entities to python
+- allow python 3.9 as baseline version
+
 ### Deprecated
 
 ### Removed